Added option to get configuration from environment variables
This feature allows the script to be used as a convenient way to get
build artifacts out of Travis CVI and into a given Dropbox account. The
DBU_* environment variables can be set securely in the user's Travis
repository settings page. The option of having a configuration file is not
convenient because the file would have to be encrypted and checked in the
repository, but then after each user has cloned the repository he/she would
have to replace the encrypted file with his/her own.